refactor(LottoController) : 컨트롤러 단에서 View와 Model 함수를 연결한다
기존 : App.js에서 모든 함수를 불러와 연동
변경 후 : Controller 단으로 App.js 함수를 옮긴 후, 기능 별로 함수를 통해 분리
* App.js를 간결하게 변경하고, Controller 단도 기능별로 분리함으로써 가독성을 높였다
