+++ src/App.js
@@ -1,33 +1,18 @@
-import InputView from "./view/InputView.js";
-import LottoNumbersModel from "./model/LottoNumbersModel.js";
-import OutputView from "./view/OutputView.js";
-import Lotto from "./model/Lotto.js";
-import CalculateProfitModel from "./model/CalculateProfitModel.js";
+import LottoController from "./controller/LottoController.js";
 
 class App {
   constructor() {
-    this.inputView = new InputView();
-    this.outputView = new OutputView();
-    this.lottoNumbersModel = new LottoNumbersModel();
-    this.calculateProfitModel = new CalculateProfitModel();
+    this.controller = new LottoController();
   }
   async run() {
-    const lottoPrice = await this.inputView.getLottoPrice();
-    const lottoCount = this.lottoNumbersModel.getCount(lottoPrice);
-    const lottoNumbers = this.lottoNumbersModel.generate(lottoCount);
-    this.outputView.printLottoNumber(lottoNumbers);
-
-    const lottoAnswer = await this.inputView.getLottoAnswer();
-    this.lotto = new Lotto(lottoAnswer);
-    const bonusAnswer = await this.inputView.getBonusAnswer(lottoAnswer);
-    const winningCount = this.lotto.CountWinningStats(
+    const { lottoPrice, lottoNumbers } =
+      await this.controller.setLottoNumbers();
+    const bonusAnswer = await this.controller.setLottoAnswers();
+    const winningCount = this.controller.countWinning(
       lottoNumbers,
       bonusAnswer
     );
-    this.outputView.printWinningCount(winningCount);
-
-    const rate = this.calculateProfitModel.getRate(winningCount, lottoPrice);
-    this.outputView.printProfitRate(rate);
+    this.controller.calculateProfit(winningCount, lottoPrice);
   }
 }
